+++ back-end/src/main/java/admin/exception/http/BadRequestException.java
@@ -1,4 +1,4 @@
-package admin.exception;
+package admin.exception.http;
 
 public class BadRequestException extends CustomException {
