+++ back-end/src/main/java/admin/exception/http/CustomException.java
@@ -1,4 +1,4 @@
-package admin.exception;
+package admin.exception.http;
 
 public class CustomException extends RuntimeException {
